lambda1/requirements.txt
python_json_logger
polygon-api-client
awswrangler